requirements.txt
Flask==1.1.2
azure-identity